backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.8.1
boto3==1.34.96
botocore==1.34.96
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.5
fastcore==1.5.33
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.4.1
ghapi==1.0.5
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.48
langchain-text-splitters==0.0.1
langsmith==0.1.52
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.2
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.29
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
flask_cors==5.0.0
anyio==4.4.0
defusedxml==0.7.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
jira==3.8.0
jiter==0.5.0
oauthlib==3.2.2
openai==1.45.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1


